+++ libraries/LTC2991/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.6] - 2023-09-23
+- add Wire1 support for ESP32
+- update readme.md
+- add defines for readability
+
+
 ## [0.1.5] - 2023-03-06
 - fix examples default address to 0x48
 - add getAddress() function